feat(output): add index and index_change
the idea here is state to show the current step for a euclid output

`Output::Gate` can be thought of as an `Output::Euclid` where the
density equals length so that every step is true. In that case prob is
the only determination of if a step is on or not.

The above means that there is no difference in tick between a gate and
an output. tick is the hot path so that is good.

The downside is that there is a bit of book keeping necessary when
switching output types, but that seems to be a good trade off.